pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kali"
version = "0.1.0"
description = "Chaos engineering toolkit — fault injection, steady-state verification, and observability integrations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["chaos-engineering", "fault-injection", "resilience", "sre"]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2",
    "httpx>=0.27",
    "pyyaml>=6",
    "anyio>=4",
    "croniter>=2",
]

[project.optional-dependencies]
prometheus = ["prometheus-client>=0.20"]
datadog = ["datadog>=0.47"]
dev = [
    "pytest>=8",
    "pytest-anyio>=0.0.0",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
kali = "kali.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/kali"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.9"
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
